requirements-conda-dev.txt
# REQUIREMENTS FOR CONDA INSTALLATION (EXCLUDING DTOCEAN PACKAGES)
configobj
pyyaml
setuptools
